+++ src/components/Filterbar/Filterbar.jsx
@@ -222,10 +222,9 @@ function Filterbar({
 
     <div className="absolute text-xs text-purple-800" style={{left: '110px', fontSize: '0.75rem', top: '16px'}}>
       versie <a href="https://github.com/Stichting-CROW/dashboarddeelmobiliteit-app/blob/main/RELEASES.md#dashboard-deelmobiliteit-app-releases" target='_blank' rel="external" className="underline">
-        2025-04-01
+        2025-04-26
       </a><br />
-      - Beter beheer multipolygonen<br />
-      - Mijksenaar logo bij elke <a href="/map/beleidshubs" className="underline">Zone</a><br />
+      Verwijder polygon uit multipolygon<br />
     </div>
   </>
 }